Remove extraneous files and move utils to correct folders
